refactor: move files to dedicated folders for better structuring
