+++ class-components/src/main.tsx
@@ -2,11 +2,11 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { BrowserRouter, Routes, Route } from 'react-router';
 import './index.css';
-import App from './components/app/App.tsx';
+import App from './pages/app/App.tsx';
 import ErrorBoundary from './components/error-boundary/error-boundary.tsx';
-import About from './pages/about-component.tsx';
-import NotFound from './pages/not-found-component.tsx';
-import Details from './pages/details.tsx';
+import About from './pages/about/about-component.tsx';
+import NotFound from './pages/not-found/not-found-component.tsx';
+import Details from './pages/details/details.tsx';
 
 const root = document.getElementById('root');
 if (root) {
