+++ class-components/src/pages/details/details.css
