+++ class-components/src/pages/details/details.tsx
